+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This is used for recording Invader's changes. This changelog is based on
 ### Added
 - invader-refactor: Added `-D` which performs the operation without making any
   actual changes
+- invader-refactor: Added `-s` which only writes to a single tag
 
 ### Changed
 - invader-build: Now errors if any tag has an animation graph but no model